fix: Organized project structure by placing all feature-related files inside the features/ directory
