+++ lib/features/gallery/presentation/bloc/image_gallery_state.dart
