+++ lib/features/gallery/presentation/widget/common_circular_loader.dart
@@ -19,7 +19,6 @@ class CommonCircularLoader extends StatelessWidget {
             ),
           ),
         ),
-
       ],
     );
   }
