+++ lib/features/image_fetcher/presentation/bloc/image_fetcher_state.dart
