+++ lib/features/image_fetcher/presentation/pages/album_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:image_fetcher/core/styles/colors.dart';
-import 'package:image_fetcher/gallery/presentation/bloc/image_gallery_cubit.dart';
-import 'package:image_fetcher/gallery/presentation/widget/common_circular_loader.dart';
-import 'package:image_fetcher/image_fetcher/presentation/bloc/image_fetcher_cubit.dart';
-import 'package:image_fetcher/image_fetcher/presentation/widget/permission__handler_widget.dart';
-import 'package:image_fetcher/gallery/presentation/pages/all_photos_view.dart';
+import 'package:image_fetcher/features/gallery/presentation/bloc/image_gallery_cubit.dart';
+import 'package:image_fetcher/features/gallery/presentation/widget/common_circular_loader.dart';
+import 'package:image_fetcher/features/image_fetcher/presentation/bloc/image_fetcher_cubit.dart';
+import 'package:image_fetcher/features/image_fetcher/presentation/widget/permission__handler_widget.dart';
+import 'package:image_fetcher/features/gallery/presentation/pages/all_photos_view.dart';
 import 'package:image_fetcher/core/dependencies/injectable.dart';
 
 class AlbumPage extends StatelessWidget {
@@ -30,7 +30,8 @@ class AlbumPage extends StatelessWidget {
       }, builder: (context, state) {
         switch (state) {
           case ImageFetcherLoadingState():
-            return const CommonCircularLoader(loaderColor: AppColors.themeColor);
+            return const CommonCircularLoader(
+                loaderColor: AppColors.themeColor);
           case ImageFetcherInitial():
           case ImageFetcherPermissionPermanentlyDenied():
           case ImageFetcherPermissionDenied():
