+++ lib/main.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:image_fetcher/core/styles/colors.dart';
-import 'package:image_fetcher/image_fetcher/presentation/bloc/image_fetcher_cubit.dart';
+import 'package:image_fetcher/features/image_fetcher/presentation/bloc/image_fetcher_cubit.dart';
 
 import 'core/dependencies/injectable.dart';
-import 'image_fetcher/presentation/pages/album_page.dart';
+import 'features/image_fetcher/presentation/pages/album_page.dart';
 
 void main() {
   configureDependencies();
